% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{B2721}
\alias{B2721}
\title{Autotetraploid potato dataset}
\format{
An object of class \code{mappoly.data} from the package \pkg{mappoly}.
}
\usage{
B2721
}
\description{
A dataset of the B2721 population which derived from a cross between 
two tetraploid potato varieties: Atlantic × B1829-5.
}
\examples{
library(mappoly)
print(B2721)
}
\references{
Mollinari M, Garcia AAF (2019) Linkage analysis and haplotype phasing in experimental autopolyploid populations with high ploidy level using hidden Markov models, \emph{G3: Genes|Genomes|Genetics} 9 (10): 3297-3314. \doi{10.1534/g3.119.400378}

    Pereira GS, Gemenet DC, Mollinari M, Olukolu BA, Wood JC, Mosquera V, Gruneberg WJ, Khan A, Buell CR, Yencho GC, Zeng ZB (2020) Multiple QTL mapping in autopolyploids: a random-effect model approach with application in a hexaploid sweetpotato full-sib population, \emph{Genetics} 215 (3): 579-595. \doi{10.1534/genetics.120.303080}.
    
    Pereira GS, Mollinari M, Schumann MJ, Clough ME, Zeng ZB, Yencho C (2021) The recombination landscape and multiple QTL mapping in a \emph{Solanum tuberosum} cv. ‘Atlantic’-derived F_1 population. Heredity. \doi{10.1038/s41437-021-00416-x}.
}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
\keyword{datasets}
