% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breeding_values.R
\name{breeding_values}
\alias{breeding_values}
\alias{plot.qtlpoly.bvalues}
\title{Prediction of QTL-based breeding values from REMIM model}
\usage{
breeding_values(data, fitted)

\method{plot}{qtlpoly.bvalues}(x, pheno.col = NULL)
}
\arguments{
\item{data}{an object of class \code{qtlpoly.data}.}

\item{fitted}{an object of class \code{qtlpoly.fitted}.}

\item{x}{an object of class \code{qtlpoly.bvalues} to be plotted.}

\item{pheno.col}{a numeric vector with the phenotype column numbers to be plotted; if \code{NULL}, all phenotypes from \code{'data'} will be included.}
}
\value{
An object of class \code{qtlpoly.bvalues} which is a list of \code{results} for each trait containing the following components:

    \item{pheno.col}{a phenotype column number.}
    \item{y.hat}{a column matrix of breeding value for each individual.}

A \pkg{ggplot2} histogram with the distribution of breeding values.
}
\description{
Computes breeding values for each genotyped individual based on multiple QTL models
}
\examples{
  \dontrun{
  # load raw data
  data(maps)
  data(pheno)

  # estimate conditional probabilities using mappoly package
  library(mappoly)
  genoprob <- lapply(maps, calc_genoprob)

  # prepare data
  data <- read_data(ploidy = 6, geno.prob = genoprob, pheno = pheno, step = 1)

  # perform remim
  remim.mod <- remim(data = data, w.size = 15, sig.fwd = 0.01, sig.bwd = 0.0001,
    d.sint = 1.5, n.clusters = 4, plot = "remim")

  # fit model
  fitted.mod <- fit_model(data = data, model = remim.mod, probs = "joint", 
    polygenes = "none")

  # predict genotypic values
  y.hat <- breeding_values(data = data, fitted = fitted.mod)
  plot(y.hat)
  }
  
}
\references{
Pereira GS, Gemenet DC, Mollinari M, Olukolu BA, Wood JC, Mosquera V, Gruneberg WJ, Khan A, Buell CR, Yencho GC, Zeng ZB (2020) Multiple QTL mapping in autopolyploids: a random-effect model approach with application in a hexaploid sweetpotato full-sib population, \emph{Genetics} 215 (3): 579-595. \url{http://doi.org/10.1534/genetics.120.303080}.
}
\seealso{
\code{\link[qtlpoly]{read_data}}, \code{\link[qtlpoly]{fit_model}}
}
\author{
Guilherme da Silva Pereira, \email{gdasilv@ncsu.edu}
}
