% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genoprob4x}
\alias{genoprob4x}
\title{Tetraploid potato genotype probabilities}
\format{
An object of class \code{"mappoly.map"} from the package \pkg{mappoly}, which is a list of three linkage groups (LGs):

\describe{
  \item{LG 1}{538 markers distributed along 112.2 cM}
  \item{LG 2}{329 markers distributed along 54.6 cM}
  \item{LG 3}{443 markers distributed along 98.2 cM}
}
}
\usage{
data(genoprob4x)
}
\description{
Genotype probabilities for three chromosomes from a tetraploid potato full-sib population (Atlantic x B1829-5).
}
\examples{
\dontrun{
data(genoprob4x)
}
}
\references{
Pereira GS, Gemenet DC, Mollinari M, Olukolu BA, Wood JC, Mosquera V, Gruneberg WJ, Khan A, Buell CR, Yencho GC, Zeng ZB (2019) Multiple QTL mapping in autopolyploids: a random-effect model approach with application in a hexaploid sweetpotato full-sib population, \emph{Genetics} 215 (3): 579-595. \url{http://doi.org/10.1534/genetics.120.303080}.
    
    Pereira GS, Mollinari M, Schumann MJ, Clough ME, Yencho C (2020) The recombination landscape and multiple QTL mapping in a \emph{Solanum tuberosum} cv. ‘Atlantic’-derived F_1 population. bioRxiv. \url{https://doi.org/10.1101/2020.08.24.265397}.
}
\author{
Guilherme da Silva Pereira, \email{gdasilv@ncsu.edu}
}
\keyword{datasets}
