% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{h}
\alias{h}
\title{Suppress annoying warning from varComp function}
\usage{
h(w)
}
\arguments{
\item{w}{a command for fitting reduced models}
}
\description{
Suppress annoying warning from varComp function
}
\keyword{internal}
