% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{maps}
\alias{maps}
\title{Simulated autohexaploid map}
\format{An object of class \code{"mappoly.map"} from the package \pkg{mappoly}, which is a list of three linkage groups (LGs):

\describe{
  \item{LG 1}{538 markers distributed along 112.2 cM}
  \item{LG 2}{329 markers distributed along 54.6 cM}
  \item{LG 3}{443 markers distributed along 98.2 cM}
}}
\usage{
data(maps)
}
\description{
A simulated map containing three homology groups of a hypotetical cross between two autohexaploid individuals.
}
\examples{
\dontrun{
data(maps)
library(mappoly)
plot(maps)
}
}
\references{
Pereira GS, Gemenet DC, Mollinari M, Olukolu BA, Wood JC, Mosquera V, Gruneberg WJ, Khan A, Buell CR, Yencho GC, Zeng ZB (2019) Multiple QTL mapping in autopolyploids: a random-effect model approach with application in a hexaploid sweetpotato full-sib population, \emph{bioRxiv}. \url{doi:}.

    Mollinari M, Garcia AAF (2019) Linkage analysis and haplotype phasing in experimental autopolyploid populations with high ploidy level using hidden Markov models, \emph{bioRxiv}. \url{https://doi.org/10.1101/415232}
}
\seealso{
\code{\link[mappoly]{hexafake}}, \code{\link[qtlpoly]{pheno}}
}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
\keyword{datasets}
