% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{mmer_adapted}
\alias{mmer_adapted}
\title{Adapt to sommer MNR core function}
\usage{
mmer_adapted(
  Y,
  X = NULL,
  Z = NULL,
  R = NULL,
  W = NULL,
  method = "NR",
  init = NULL,
  iters = 20,
  tolpar = 0.001,
  tolparinv = 1e-06,
  draw = FALSE,
  silent = FALSE,
  constraint = TRUE,
  EIGEND = FALSE,
  forced = NULL,
  IMP = FALSE,
  complete = TRUE,
  check.model = TRUE,
  restrained = NULL,
  REML = TRUE,
  init.equal = TRUE,
  date.warning = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{void}{internal function to be documented}
}
\description{
Adapts genomic incidence and relationship (varcov) matrices to run using sommer's C++ core function (v. 4.0 or higher)
Function adapted from sommer v. 3.6 (Author: Giovanny Covarrubias-Pazaran)
}
\references{
Covarrubias-Pazaran G (2016) Genome-assisted prediction of quantitative traits using the R package sommer. \emph{PLoS ONE} 11 (6): 1–15. \doi{10.1371/journal.pone.0156744}.

    Pereira GS, Gemenet DC, Mollinari M, Olukolu BA, Wood JC, Mosquera V, Gruneberg WJ, Khan A, Buell CR, Yencho GC, Zeng ZB (2020) Multiple QTL mapping in autopolyploids: a random-effect model approach with application in a hexaploid sweetpotato full-sib population, \emph{Genetics} 215 (3): 579-595. \doi{10.1534/genetics.120.303080}.
}
\author{
Gabriel de Siqueira Gesteira, \email{gdesiqu@ncsu.edu}
}
\keyword{internal}
