% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_qtl.R
\name{optimize_qtl}
\alias{optimize_qtl}
\alias{print.qtlpoly.optimize}
\title{Model optimization}
\usage{
optimize_qtl(data, model, sig.bwd = 1e-04, polygenes = FALSE,
  n.clusters = NULL, plot = "optimize", verbose = TRUE)

\method{print}{qtlpoly.optimize}(x, pheno.col = NULL)
}
\arguments{
\item{data}{an object of class \code{qtlpoly.data}.}

\item{model}{an object of class \code{qtlpoly.model} containing the QTL to be optimized.}

\item{sig.bwd}{the desired score-based \emph{p}-value threshold for backward elimination, e.g. 0.0001 (default).}

\item{polygenes}{if \code{TRUE} all QTL but the one being tested are treated as a single polygenic effect, if \code{FALSE} (default) all QTL effect variances have to estimated.}

\item{n.clusters}{number of parallel processes to spawn.}

\item{plot}{a suffix for the file's name containing plots of every QTL optimization round, e.g. "optimize" (default); if \code{NULL}, no file is produced.}

\item{verbose}{if \code{TRUE} (default), current progress is shown; if \code{FALSE}, no output is produced.}

\item{x}{an object of class \code{qtlpoly.optimize} to be printed.}

\item{pheno.col}{a numeric vector with the phenotype columns to be printed; if \code{NULL}, all phenotypes from \code{'data'} will be included.}
}
\value{
An object of class \code{qtlpoly.optimize} which contains a list of \code{results} for each trait with the following components:

    \item{pheno.col}{a phenotype column number.}
    \item{stat}{a vector containing values from score statistics.}
    \item{pval}{a vector containing \emph{p}-values from score statistics.}
    \item{qtls}{a data frame with information from the mapped QTL.}
}
\description{
Tests each QTL at a time and updates its position (if it changes) or drops the QTL (if non-significant).
}
\examples{
  \dontrun{
  # load raw data
  data(maps)
  data(pheno)

  # estimate conditional probabilities using 'mappoly' package
  library(mappoly)
  genoprob <- lapply(maps, calc_genoprob)

  # prepare data
  data <- read_data(ploidy = 6, geno.prob = genoprob, pheno = pheno, step = 1)

  # build null models
  null.mod <- null_model(data = data, n.clusters = 4, plot = "null")

  # perform forward search
  search.mod <- search_qtl(data = data, model = null.mod, w.size = 15, sig.fwd = 0.01,
    n.clusters = 4, plot = "search")

  # optimize model
  optimize.mod <- optimize_qtl(data = data, model = search.mod, sig.bwd = 0.0001,
    n.clusters = 4, plot = "optimize")
  }

}
\references{
Pereira GS, Gemenet DC, Mollinari M, Olukolu BA, Wood JC, Mosquera V, Gruneberg WJ, Khan A, Buell CR, Yencho GC, Zeng ZB (2019) Multiple QTL mapping in autopolyploids: a random-effect model approach with application in a hexaploid sweetpotato full-sib population, \emph{bioRxiv}. \url{doi:}.
    
    Qu L, Guennel T, Marshall SL (2013) Linear score tests for variance components in linear mixed models and applications to genetic association studies. \emph{Biometrics} 69 (4): 883–92. \url{doi:10.1111/biom.12095}.
}
\seealso{
\code{\link[qtlpoly]{read_data}}, \code{\link[qtlpoly]{null_model}}, \code{\link[qtlpoly]{search_qtl}}
}
\author{
Guilherme da Silva Pereira, \email{gdasilv@ncsu.edu}
}
