% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutations.R
\name{permuta}
\alias{permuta}
\title{Compute LOD score from a simulated phenotype using fixed-effect interval mapping (FEIM) model}
\usage{
permuta(
  x,
  data = data,
  offset.data = offset.data,
  pheno.col = pheno.col,
  p = p
)
}
\arguments{
\item{x}{an array with the number of simulations}

\item{data}{an object of 'qtlpoly.data'}

\item{pheno.col}{a vector of phenotype column numbers}

\item{p}{the phenotype column number to be evaluated}
}
\description{
Compute LOD score from a simulated phenotype using fixed-effect interval mapping (FEIM) model
}
\keyword{internal}
