% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_qtl.R
\name{plot_qtl}
\alias{plot_qtl}
\title{QTL heritability and significance plot}
\usage{
plot_qtl(data = data, model = model, fitted = fitted,
  pheno.col = NULL, main = NULL, drop = FALSE)
}
\arguments{
\item{data}{an object of class \code{qtlpoly.data}.}

\item{model}{an object of class \code{qtlpoly.profile} or \code{qtlpoly.remim}.}

\item{fitted}{an object of class \code{qtlpoly.fitted}.}

\item{pheno.col}{the desired phenotype column numbers to be plotted. The order here specifies the order of plotting (from top to bottom.)}

\item{main}{plot title; if \code{NULL} (the default), no title is shown.}

\item{drop}{if \code{FALSE}, shows the names of all traits from \code{pheno.col}, even of those with no QTLs; if \code{TRUE} (the default), shows only the traits with QTL(s).}
}
\value{
A \pkg{ggplot2} with dots representing the QTLs.
}
\description{
Creates a plot where dot sizes and colors represent the QTLs heritabilities and their \emph{p}-values, respectively.
}
\examples{
  \dontrun{
  # load raw data
  data(maps)
  data(pheno)

  # estimate conditional probabilities using mappoly package
  library(mappoly)
  genoprob <- lapply(maps, calc_genoprob)

  # prepare data
  data <- read_data(ploidy = 6, geno.prob = genoprob, pheno = pheno, step = 1)

  # perform remim
  remim.mod <- remim(data = data, w.size = 15, sig.fwd = 0.01, sig.bwd = 0.0001,
    d.sint = 1.5, n.clusters = 4, plot = "remim")

  # fit model
  fitted.mod <- fit_model(data=data, model=remim.mod, probs="joint", polygenes="none")

  # plot qtls
  plot_qtl(data = data, model = remim.mod, fitted = fitted.mod)
  }
}
\references{
Pereira GS, Gemenet DC, Mollinari M, Olukolu BA, Wood JC, Mosquera V, Gruneberg WJ, Khan A, Buell CR, Yencho GC, Zeng ZB (2019) Multiple QTL mapping in autopolyploids: a random-effect model approach with application in a hexaploid sweetpotato full-sib population, \emph{bioRxiv}. \url{doi:}.
}
\seealso{
\code{\link[qtlpoly]{read_data}}, \code{\link[qtlpoly]{remim}}, \code{\link[qtlpoly]{fit_model}}
}
\author{
Guilherme da Silva Pereira, \email{gdasilv@ncsu.edu}
}
