% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling.R
\name{resampling}
\alias{resampling}
\alias{print.qtlpoly.rsmp}
\alias{plot.qtlpoly.rsmp}
\title{Random-effect multiple interval mapping (REMIM) model score-based resampling method}
\usage{
resampling(
  data,
  n.sim = 1000,
  alpha = c(0.2, 0.05),
  n.clusters = NULL,
  seed = 123,
  verbose = TRUE
)

\method{print}{qtlpoly.rsmp}(x, pheno.col = NULL, probs = c(0.9, 0.95))

\method{plot}{qtlpoly.rsmp}(x, alpha = c(0.2, 0.05))
}
\arguments{
\item{data}{an object of class \code{qtlpoly.data}.}

\item{n.sim}{a number of simulations, e.g. 1000 (default).}

\item{n.clusters}{a number of parallel processes to spawn.}

\item{seed}{an integer for the \code{set.seed()} function; if \code{NULL}, no reproducible seeds are set.}

\item{verbose}{if \code{TRUE} (default), current progress is shown; if \code{FALSE}, no output is produced.}

\item{x}{an object of class \code{qtlpoly.perm} to be printed or plotted.}

\item{pheno.col}{a numeric vector with the phenotype columns to be analyzed; if \code{NULL} (default), all phenotypes from \code{'data'} will be included.}

\item{probs}{a vector of probability values in [0, 1] representing the quantiles, e.g. c(0.90, 0.95) for the 90\% and 95\% quantiles.}
}
\value{
An object of class \code{qtlpoly.perm} which contains a list of \code{results} for each trait with the maximum LOD score per permutation.

LOD score thresholds for given quantiles for each trait.

A \pkg{ggplot2} histogram with the distribution of ordered maximum LOD scores and thresholds for given quantiles for each trait.
}
\description{
Stores P-values from maximum score statistics for a number of resampling of given linkage map.
}
\examples{
  \dontrun{
  # load raw data
  data(maps)
  data(pheno)

  # estimate conditional probabilities using mappoly package
  library(mappoly)
  genoprob <- lapply(maps, calc_genoprob)

  # prepare data
  data <- read_data(ploidy = 6, geno.prob = genoprob, pheno = pheno, step = 1)

  # perform resampling
  score.rsmp <- resampling(data = data, n.sim = 1000, n.clusters = 4)
  }

}
\references{
Zou F, Fine JP, Hu J, Lin DY (2004) An efficient resampling method for assessing genome-wide statistical significance in mapping quantitative trait loci, \emph{Genetics} 168 (4): 2307-2316. \url{https://doi.org/10.1534/genetics.104.031427}.
    Pereira GS, Gemenet DC, Mollinari M, Olukolu BA, Wood JC, Mosquera V, Gruneberg WJ, Khan A, Buell CR, Yencho GC, Zeng ZB (2020) Multiple QTL mapping in autopolyploids: a random-effect model approach with application in a hexaploid sweetpotato full-sib population, \emph{Genetics} 215 (3): 579-595. \url{http://doi.org/10.1534/genetics.120.303080}.
}
\seealso{
\code{\link[qtlpoly]{feim}}
}
\author{
Guilherme da Silva Pereira, \email{gdasilv@ncsu.edu}
}
