% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_qtl.R
\name{search_qtl}
\alias{search_qtl}
\alias{print.qtlpoly.search}
\title{QTL forward search}
\usage{
search_qtl(data, model, w.size = 15, sig.fwd = 0.01,
  polygenes = FALSE, n.rounds = Inf, n.clusters = NULL,
  plot = "search", verbose = TRUE)

\method{print}{qtlpoly.search}(x, pheno.col = NULL)
}
\arguments{
\item{data}{an object of class \code{qtlpoly.data}.}

\item{model}{an object of class \code{qtlpoly.model} from which a forward search will start.}

\item{w.size}{the window size (in cM) to avoid on either side of QTL already in the model when looking for a new QTL.}

\item{sig.fwd}{the desired score-based \emph{p}-value threshold for forward search, e.g. 0.01 (default).}

\item{polygenes}{if \code{TRUE} all QTL but the one being tested are treated as a single polygenic effect; if \code{FALSE} (default) all QTL effect variances have to estimated.}

\item{n.rounds}{number of search rounds; if \code{Inf} (default) forward search will stop when no more significant positions can be found.}

\item{n.clusters}{number of parallel processes to spawn.}

\item{plot}{a suffix for the file's name containing plots of every QTL search round, e.g. "search" (default); if \code{NULL}, no file is produced.}

\item{verbose}{if \code{TRUE} (default), current progress is shown; if \code{FALSE}, no output is produced.}

\item{x}{an object of class \code{qtlpoly.search} to be printed.}

\item{pheno.col}{a numeric vector with the phenotype column numbers to be printed; if \code{NULL}, all phenotypes from \code{'data'} will be included.}
}
\value{
An object of class \code{qtlpoly.search} which contains a list of \code{results} for each trait with the following components:

    \item{pheno.col}{a phenotype column number.}
    \item{stat}{a vector containing values from score statistics.}
    \item{pval}{a vector containing \emph{p}-values from score statistics.}
    \item{qtls}{a data frame with information from the mapped QTL.}
}
\description{
Searches for QTL and adds them one at a time to a multiple random-effect QTL model based on score statistics.
}
\examples{
  \dontrun{
  # load raw data
  data(maps)
  data(pheno)

  # estimate conditional probabilities using 'mappoly' package
  library(mappoly)
  genoprob <- lapply(maps, calc_genoprob)

  # prepare data
  data <- read_data(ploidy = 6, geno.prob = genoprob, pheno = pheno, step = 1)

  # build null models
  null.mod <- null_model(data = data, n.clusters = 4, plot = "null")

  # perform forward search
  search.mod <- search(data = data, model = null.mod, w.size = 15, sig.fwd = 0.01,
    n.clusters = 4, plot = "search")
  }

}
\references{
Pereira GS, Gemenet DC, Mollinari M, Olukolu BA, Wood JC, Mosquera V, Gruneberg WJ, Khan A, Buell CR, Yencho GC, Zeng ZB (2019) Multiple QTL mapping in autopolyploids: a random-effect model approach with application in a hexaploid sweetpotato full-sib population, \emph{bioRxiv}. \url{doi:}.
    
    Qu L, Guennel T, Marshall SL (2013) Linear score tests for variance components in linear mixed models and applications to genetic association studies. \emph{Biometrics} 69 (4): 883–92. \url{doi:10.1111/biom.12095}.
}
\seealso{
\code{\link[qtlpoly]{read_data}}, \code{\link[qtlpoly]{null_model}}
}
\author{
Guilherme da Silva Pereira, \email{gdasilv@ncsu.edu}
}
