% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_qtl.R
\name{simulate_qtl}
\alias{simulate_qtl}
\alias{print.qtlpoly.simul}
\title{Simulations of multiple QTL}
\usage{
simulate_qtl(data, mu = 0, h2.qtl = c(0.3, 0.2, 0.1), var.error = 1,
  n.sim = 1000, w.size = 20, seed = 123, verbose = TRUE)

\method{print}{qtlpoly.simul}(x, detailed = FALSE)
}
\arguments{
\item{data}{an object of class \code{qtlpoly.data}.}

\item{mu}{simulated phenotype mean, e.g. 0 (default).}

\item{h2.qtl}{vector with QTL heritabilities, e.g. c(0.3, 0.2, 0.1) for three QTL (default).}

\item{var.error}{simulated error variance, e.g. 1 (default).}

\item{n.sim}{number of simulations, e.g. 1000 (default).}

\item{w.size}{the window size (in centiMorgans) to avoid on either side of QTL already in the model when looking selecting new QTL, e.g. 20 (default).}

\item{seed}{integer for the \code{set.seed()} function.}

\item{verbose}{if \code{TRUE} (default), current progress is shown; if \code{FALSE}, no output is produced.}

\item{x}{an object of class \code{qtlpoly.sim} to be printed.}

\item{detailed}{if \code{TRUE}, detailed information on linkage groups and phenotypes in shown; if \code{FALSE}, no details are printed.}
}
\value{
An object of class \code{qtlpoly.sim} which contains a list of \code{results} with the same structure of class \code{qtlpoly.data}.
}
\description{
Simulate new phenotypes with a given number of QTL and creates new object with the same structure of class \code{qtlpoly.data} from an existing genetic map.
}
\examples{
  \dontrun{
  # load raw data
  data(maps)
  data(pheno)

  # estimate conditional probabilities using mappoly package
  library(mappoly)
  genoprob <- lapply(maps, calc_genoprob)

  # prepare data
  data <- read_data(ploidy = 6, geno.prob = genoprob, pheno = pheno, step = 1)

  # simulate new phenotypes
  sim.dat <- simulate_qtl(data = data, n.sim = 1000)
  }

}
\references{
Pereira GS, Gemenet DC, Mollinari M, Olukolu BA, Wood JC, Mosquera V, Gruneberg WJ, Khan A, Buell CR, Yencho GC, Zeng ZB (2019) Multiple QTL mapping in autopolyploids: a random-effect model approach with application in a hexaploid sweetpotato full-sib population, \emph{bioRxiv}. \url{doi:}.
}
\seealso{
\code{\link[qtlpoly]{read_data}}
}
\author{
Guilherme da Silva Pereira, \email{gdasilv@ncsu.edu}
}
