% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pheno4x}
\alias{pheno4x}
\title{Autotetraploid potato phenotypes}
\format{
A data frame of phenotypes with 156 named individuals in rows and three named phenotypes in columns, which are:

\describe{
  \item{FM07}{Foliage maturity evaluated in 2007.}
  \item{FM08}{Foliage maturity evaluated in 2008.}
  \item{FM14}{Foliage maturity evaluated in 2014.}
}
}
\usage{
pheno4x
}
\description{
A subset of phenotypes from a tetraploid potato full-sib family (Atlantic x B1829-5).
}
\examples{
head(pheno4x)
}
\references{
Pereira GS, Gemenet DC, Mollinari M, Olukolu BA, Wood JC, Mosquera V, Gruneberg WJ, Khan A, Buell CR, Yencho GC, Zeng ZB (2020) Multiple QTL mapping in autopolyploids: a random-effect model approach with application in a hexaploid sweetpotato full-sib population, \emph{Genetics} 215 (3): 579-595. \doi{10.1534/genetics.120.303080}.
    
    Pereira GS, Mollinari M, Schumann MJ, Clough ME, Zeng ZB, Yencho C (2021) The recombination landscape and multiple QTL mapping in a \emph{Solanum tuberosum} cv. ‘Atlantic’-derived F_1 population. Heredity. \doi{10.1038/s41437-021-00416-x}.
}
\author{
Guilherme da Silva Pereira, \email{gdasilv@ncsu.edu}
}
\keyword{datasets}
